perf: improve performance of cloning requests
An `Operation` object created by `Request::operation()` refers to
objects inside the `Request` object.  That prevents us to move the
`Request` object without cloning inside a function to handle the
`Operation` object.  However, cloning the `Request` object is
inefficient operation because the data buffer `arg` is represented
with a `Vec` which clones the data even when the data buffer is
immutable.

This commit changes the type of the data buffer to `bytes::Bytes`.  It
can clone efficiently.

This is the reason why we renamed `Bytes` to `AtomicBytes` in the
previous commit.